build: take both arrangements to zero warnings, and re-measure the count
The record said eleven warnings remained, across three files, and offered
that as the basis for a possible warnings gate. Re-derived with a clean
build of each arrangement, because it is a scope number and nothing
downstream re-derives one:

    -Os, GUI and CLI on          0
    sanitized -Og               14

Fourteen across eight files, and six of them in shipped library sources
the list does not mention -- the `bit_get` family in spool/plan.c,
spool/spool.c, spool/scrub.c and chunk/reassembly.c, plus qr/qr.c's two.

A warning is only emitted by a file that gets recompiled, so every count
taken from a `make check` log is a count over whatever that run happened
to rebuild. The gate log this was first checked against showed ten
sites; a clean build of the same tree shows fourteen. Neither run was
wrong and only one of them answers how many warnings the tree has, so
the method is recorded beside the number.

The six library ones are one idiom: a uint8_t or uint16_t promotes to
int, and `& 1u` then converts that signed value. Benign in every case --
the operand cannot be negative and the mask yields 0 or 1 -- so the fix
makes the promotion explicit and changes no value.

Six of the eight in tests were the same shape or a shadowed local. Two
were worth more than the warning. reassembly_fuzz.c cast both arms of a
conditional to uint16_t and the conditional promoted them straight back
to int, so the assignment converted anyway; the casts were saying the
right thing in a place where C does not listen. And spool_file_test.c
built a path with snprintf without reading the result: it cannot
overflow, but it can truncate, at which point the case opens a different
file and passes having exercised a spool it did not mean.

The four shadowed locals were renamed with a tool that walks to the
enclosing block's closing brace and refuses to touch anything outside
it. The outer name has the same spelling by definition, so a file-wide
rewrite would have renamed the thing being shadowed.

Value preservation is checked rather than argued: manifest_fuzz,
reassembly_fuzz and spool_fuzz each report tallies over 20000 cases that
are identical before and after, and the three suites whose sources were
renamed in report identical counts.

Both arrangements measure zero now. Whether to add a gate is still not
this change's call, but the answer no longer has to start from a waiver
list.

+++ chain/test/chain_guided.c
@@ -108,7 +108,7 @@ struct stub {
 
 static int stub_signature_is_good(uint16_t good, const uint8_t *pubkey)
 {
-	return (int)((good >> (pubkey[0] % 16u)) & 1u);
+	return (int)(((unsigned)good >> (pubkey[0] % 16u)) & 1u);
 }
 
 static int stub_verify(void *ctx, const uint8_t pubkey[FZN_PUBKEY_LEN], const uint8_t *msg,
